+++ src/core/client.ts
@@ -11,7 +11,8 @@ import {
   filter,
   filterUrl,
   formatCGIData,
-  getSourceForBacktrace
+  getSourceForBacktrace,
+  runAfterNotifyHandlers
 } from './util'
 import {
   Config, Logger, BreadcrumbRecord, BeforeNotifyHandler, AfterNotifyHandler, Notice, Noticeable
@@ -137,31 +138,42 @@ export default class Client {
   }
 
   notify(noticeable: Noticeable, name: string | Partial<Notice> = undefined, extra: Partial<Notice> = undefined): boolean {
-    if (this.config.disabled) {
-      this.logger.warn('Deprecation warning: instead of `disabled: true`, use `reportData: false` to explicitly disable Honeybadger reporting. (Dropping notice: honeybadger.js is disabled)')
-      return false
-    }
 
-    if (!this.__reportData()) {
-      this.logger.debug('Dropping notice: honeybadger.js is in development mode')
-      return false
-    }
-
-    if (!this.config.apiKey) {
-      this.logger.warn('Unable to send error report: no API key has been configured')
-      return false
-    }
-
+    let preConditionError: Error = null
     const notice = this.makeNotice(noticeable, name, extra)
     if (!notice) {
-      return false
+      preConditionError = new Error('Could not make Notice')
+    }
+
+    if (!preConditionError && this.config.disabled) {
+      const msg = 'Deprecation warning: instead of `disabled: true`, use `reportData: false` to explicitly disable Honeybadger reporting. (Dropping notice: honeybadger.js is disabled)'
+      this.logger.warn(msg)
+      preConditionError = new Error(msg)
+    }
+
+    if (!preConditionError && !this.__reportData()) {
+      const msg = 'Dropping notice: honeybadger.js is in development mode'
+      this.logger.debug(msg)
+      preConditionError = new Error(msg)
+    }
+
+    if (!preConditionError && !this.config.apiKey) {
+      const msg = 'Unable to send error report: no API key has been configured'
+      this.logger.warn(msg)
+      preConditionError = new Error(msg)
     }
 
     // we need to have the source file data before the beforeNotifyHandlers,
     // in case they modify them
-    const sourceCodeData = notice.backtrace.slice()
+    const sourceCodeData = notice ? notice.backtrace.slice() : null
 
-    if (!runBeforeNotifyHandlers(notice, this.__beforeNotifyHandlers)) {
+    const beforeNotifyResult = runBeforeNotifyHandlers(notice, this.__beforeNotifyHandlers)
+    if (!preConditionError && !beforeNotifyResult) {
+      preConditionError = new Error('Will not send error report, beforeNotify handlers returned false')
+    }
+
+    if (preConditionError) {
+      runAfterNotifyHandlers(notice, this.__afterNotifyHandlers, preConditionError)
       return false
     }
 
@@ -187,7 +199,7 @@ export default class Client {
     return true
   }
 
-  protected makeNotice(noticeable: Noticeable, name: string | Partial<Notice> = undefined, extra: Partial<Notice> = undefined): Notice {
+  protected makeNotice(noticeable: Noticeable, name: string | Partial<Notice> = undefined, extra: Partial<Notice> = undefined): Notice | null {
     let notice = makeNotice(noticeable)
 
     if (name && !(typeof name === 'object')) {